fix: copy-paste error caused misparsing of double configurations

+++ vane-core/src/main/java/org/oddlama/vane/core/config/ConfigDoubleField.java
@@ -58,7 +58,7 @@ public class ConfigDoubleField extends ConfigField<Double> {
 	}
 
 	public double load_from_yaml(YamlConfiguration yaml) {
-		return yaml.getInt(yaml_path());
+		return yaml.getDouble(yaml_path());
 	}
 
 	public void load(YamlConfiguration yaml) {
